+++ template-sample-app-2/template/src/components/layers/index.js
@@ -3,6 +3,7 @@ import StoresLayer from './StoresLayer';
 import GeocoderLayer from './GeocoderLayer';
 import IsochroneLayer from './IsochroneLayer';
 import TilesetLayer from './TilesetLayer';
+import CollisionsLayer from './CollisionsLayer';
 // [hygen] Import layers
 
 export const getLayers = () => {
@@ -12,6 +13,7 @@ export const getLayers = () => {
     KpiLayer(),
     GeocoderLayer(),
     TilesetLayer(),
+    CollisionsLayer(),
     // [hygen] Add layer
   ];
 };